+++ lunaticfamily-common/src/main/java/de/janschuri/lunaticfamily/common/handler/familytree/RootAdvancement.java
@@ -7,7 +7,7 @@ import java.util.UUID;
 public class RootAdvancement extends RelationAdvancement {
     private final String background;
 
-    public RootAdvancement(String key, String background, TreeAdvancement parentKey, int id, UUID uuid, String gender, String title, String description, String skinUrl, Side side) {
+    public RootAdvancement(String key, String background, TreeAdvancement parentKey, long id, UUID uuid, String gender, String title, String description, String skinUrl, Side side) {
         super(key, parentKey, id, uuid, gender, title, description, skinUrl, 0, 0, side);
         this.background = background;
     }
